Fixed signIn & started hide/show
I fixed the auth sign in and now the code is saving the token
there was a typo in the events.js folder for onSignInSuccess.
I also added hidden and show classes to my html and added the jquery
method of hide to app.js to hide the gameboard, signout, change pass,
and play game.